Refactor: reorder gettext initialization
Refactor config file loading and locale path fetching
to read user settings prior to gettext initialization.

Factor out command line arg. parsing for readability.
Move gettext initialization function to new module.

Split out the default-config definition and the loading
of user settings from main.main() to its own module.

Move the gettext initialization to just before
the application is initialized.
